Moving sources from master to akarapetyan branch
